requirements.txt
-r requirements_core.txt
boto3
botocore
cupy-cuda11x
highdicom
imageio[pyav]
ipython
jsonargparse[signatures]
matplotlib
moviepy  # required for torch.utils.tensorboard.writer.SummaryWriter.add_video
notebook
openpyxl
optuna==3.1
pandas
pillow
plotly  # required for optuna.visualization
pydicom
pygrappa
pymapvbvd
pytorch-lightning==2.0.*
pyyaml
scikit-image
sigpy
tensorboard
tqdm
